# Ensure that `fn config_path` returns canonicalized paths.