This commit ensures that `fn config_path` in `rustfmt/src/config/mod.rs` always returns canonicalized paths.  This is achieved by canonicalizing both: 1) paths received via `CliOptions` (after checking if the path exists) as well as 2) paths returned via `get_toml_path` (canonicalizing within `fn get_toml_path`).

Fixes https://github.com/rust-lang/rustfmt/issues/6178