@ytmimi 2024-08-12
Thanks for the PR. Would you mind double checking if there's any difference in the output when running `rustfmt -v path/to/file.rs` on windows before and after this change if the paths are canonicalized. I think we might already be canonicalizing them before we display the paths, but I want to double check.

Output should look something like this:
```
Using rustfmt config file /path/to/rustfmt.toml for /path/to/file.rs
Formatting /path/to/file.rs
Spent 0.000 secs in the parsing phase, and 0.001 secs in the formatting phase
```