@anforowicz 2024-08-12
> Would you mind double checking if there's any difference in the output when running `rustfmt -v path/to/file.rs` on windows before and after this change if the paths are canonicalized. I think we might already be canonicalizing them before we display the paths, but I want to double check.

There are no output changes in this scenario.

Before the changes:

```
C:\src\rustfmt>git checkout origin/master
C:\src\rustfmt>cargo run --bin rustfmt -- -v path/to/file.rs
   Compiling rustfmt-nightly v1.7.1 (C:\src\rustfmt)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 9.50s
     Running `target\debug\rustfmt.exe -v path/to/file.rs`
Using rustfmt config file \\?\C:\src\rustfmt\rustfmt.toml for \\?\C:\src\rustfmt\path\to\file.rs
Formatting \\?\C:\src\rustfmt\path\to\file.rs
Spent 0.012 secs in the parsing phase, and 0.002 secs in the formatting phase
```

After the changes:

```
C:\src\rustfmt>git checkout path-canon
C:\src\rustfmt>cargo run --bin rustfmt -- -v path/to/file.rs
   Compiling rustfmt-nightly v1.7.1 (C:\src\rustfmt)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 8.84s
     Running `target\debug\rustfmt.exe -v path/to/file.rs`
Using rustfmt config file \\?\C:\src\rustfmt\rustfmt.toml for \\?\C:\src\rustfmt\path\to\file.rs
Formatting \\?\C:\src\rustfmt\path\to\file.rs
Spent 0.012 secs in the parsing phase, and 0.001 secs in the formatting phase
```

--------------------

As expected the path passed to `--config-path` changes (gets canonicalized) after the changes.

Before the changes:

```
C:\src\rustfmt>git checkout origin/master
C:\src\rustfmt>cargo run --bin rustfmt -- -v --config-path=c:\test\rustfmt.toml c:\test\bar.rs
   Compiling rustfmt-nightly v1.7.1 (C:\src\rustfmt)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 8.52s
     Running `target\debug\rustfmt.exe -v --config-path=c:\test\rustfmt.toml c:\test\bar.rs`
Using rustfmt config file c:\test\rustfmt.toml
thread 'main' panicked at C:\Users\lukasza\.cargo\registry\src\index.crates.io-6f17d22bba15001f\ignore-0.4.18\src\gitignore.rs:227:9:
path is expected to be under the root
...
```

After the changes:

```
C:\src\rustfmt>git checkout path-canon
C:\src\rustfmt>cargo run --bin rustfmt -- -v --config-path=c:\test\rustfmt.toml c:\test\bar.rs
   Compiling rustfmt-nightly v1.7.1 (C:\src\rustfmt)
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 8.44s
     Running `target\debug\rustfmt.exe -v --config-path=c:\test\rustfmt.toml c:\test\bar.rs`
Using rustfmt config file \\?\C:\test\rustfmt.toml
Formatting \\?\C:\test\bar.rs
Spent 0.010 secs in the parsing phase, and 0.005 secs in the formatting phase
```